Day 11 fixes
Dividing by the LCM of the tests for part 1 doesn't work unless
you happen to have a monkey that tests on 3 (because we have
an explicit divide by 3 involved). Just don't bother doing it.

Also replace LCM with a straight up multiplication as it turns
out all the tests are boringly prime.